+++ app/src/main/java/it/unicam/cs/api/parser/CommandAction.java
@@ -1,4 +1,4 @@
-package it.unicam.cs.api.parser.api;
+package it.unicam.cs.api.parser;
 
 
 /**
