+++ app/src/main/java/it/unicam/cs/api/parser/Information.java
@@ -1,4 +1,4 @@
-package it.unicam.cs.api.parser.api;
+package it.unicam.cs.api.parser;
 
 import it.unicam.cs.api.parser.parsers.AbstractParser;
