+++ app/src/main/java/it/unicam/cs/api/parser/Interpretable.java
@@ -1,4 +1,4 @@
-package it.unicam.cs.api.parser.api;
+package it.unicam.cs.api.parser;
 
 import it.unicam.cs.api.exceptions.parser.NoActionFoundException;
